Character and Enemy attack
Character:
Fire and Sap damage the enemy both with different damage amounts.
-Fire: 50 Damage
-Sap: 20 Damage applied multiple times while enemy remains inside the ring.
Enemy:
Throws a projectile that damages the player. The projectile remains in place for few seconds before it dies. If the player walks into the projectile it will damage him.
Did a small change so the first enemy doesnt spawn in the main room.
If the player reaches 0 health nothing happens so far but function has been placed with a comment TO DO: