Player Attack animations
Animations added for the Warrior and Mage with left and right mouse click.
Warrior triple attack combo stills need correcting for timings but works ok
Mage fire spell needs adjusting to fire out correctly with attack animation.
The Sap can be changed to special attack so it matches with the left click animation.